Move functions adapted from odc.aws.inventory from utils into inventory module
