+++ deafrica/monitoring/landsat_gap_report.py
@@ -24,11 +24,11 @@ from urlpath import URL
 
 from deafrica import __version__
 from deafrica.click_options import slack_url, update_stac
+from deafrica.inventory import list_inventory
 from deafrica.logs import setup_logging
 from deafrica.utils import (
     convert_str_to_date,
     download_file_to_tmp,
-    list_inventory,
     send_slack_notification,
     time_process,
 )
