+++ deafrica/monitoring/s2_c1_gap_report.py
@@ -11,9 +11,9 @@ from urlpath import URL
 
 from deafrica import __version__
 from deafrica.click_options import slack_url, update_stac
+from deafrica.inventory import list_inventory
 from deafrica.logs import setup_logging
 from deafrica.utils import (
-    list_inventory,
     send_slack_notification,
 )
